refactor: 表格列选择器逻辑优化
