+++ frontend/src/App.vue
@@ -6,15 +6,13 @@
 </template>
 
 <script lang="ts" setup>
-  import { computed, onMounted } from 'vue';
+  import { computed } from 'vue';
   import enUS from '@arco-design/web-vue/es/locale/lang/en-us';
   import zhCN from '@arco-design/web-vue/es/locale/lang/zh-cn';
   import GlobalSetting from '@/components/pure/global-setting/index.vue';
   import useLocale from '@/locale/useLocale';
-  import { useTableStore } from './store';
 
   const { currentLocale } = useLocale();
-  const tableStore = useTableStore();
   const locale = computed(() => {
     switch (currentLocale.value) {
       case 'zh-CN':
@@ -24,8 +22,5 @@
       default:
         return zhCN;
     }
-  });
-  onMounted(() => {
-    tableStore.initColumn();
   });
 </script>
